feat(server):refactor gpt and resource dependencies
