+++ packages/amplication-prisma-db/prisma/schema.prisma
@@ -604,7 +604,7 @@ enum EnumGitOrganizationType {
 
 enum EnumUserActionType {
   DBSchemaImport
-  BreakTheMonolith
+  GptConversation
 }
 
 enum CodeGeneratorVersionStrategy {
