+++ packages/amplication-server/src/core/resource/resourceBtm.types.ts
@@ -15,3 +15,23 @@ export interface EntityPartial
 export interface ResourcePartial extends Pick<Resource, "id" | "name"> {
   entities: EntityPartial[];
 }
+
+interface DataModel {
+  name: string;
+  fields: {
+    name: string;
+    dataType: string;
+  }[];
+}
+
+export interface BreakTheMonolithPromptInput {
+  dataModels: DataModel[];
+}
+
+export interface BreakTheMonolithPromptOutput {
+  microservices: {
+    name: string;
+    functionality: string;
+    dataModels: string[];
+  }[];
+}
