--- packages/amplication-server/src/core/resource/resourceBtmPrompt.types.ts
@@ -1,19 +1,0 @@
-interface DataModel {
-  name: string;
-  fields: {
-    name: string;
-    dataType: string;
-  }[];
-}
-
-export interface BreakTheMonolithPromptInput {
-  dataModels: DataModel[];
-}
-
-export interface BreakTheMonolithPromptOutput {
-  microservices: {
-    name: string;
-    functionality: string;
-    dataModels: string[];
-  }[];
-}
